+++ Imperium/src/Interface/ImperiumUI/ImperiumWindow.cs
@@ -160,7 +160,7 @@ internal abstract class ImperiumWindow : MonoBehaviour, IDragHandler, IBeginDrag
 
     protected void RegisterWidget<T>(Transform container, string path) where T : ImpWidget
     {
-        container.Find(path).gameObject.AddComponent<T>().Init(theme, tooltip, ref onOpen, ref onClose);
+        container.Find(path).gameObject.AddComponent<T>().Init(theme, tooltip, ref onOpen, ref onClose, parent);
     }
 
     /// <summary>