+++ Imperium/src/Interface/ImperiumUI/Windows/Teleport/Widgets/Teleportation.cs
@@ -1,10 +1,7 @@
-#region
-
 using System;
 using System.Collections;
 using System.Collections.Generic;
 using Imperium.Interface.Common;
-using Imperium.Interface.ImperiumUI.Windows.Teleportation.Widgets;
 using Imperium.Types;
 using Imperium.Util.Binding;
 using TMPro;
@@ -12,11 +9,9 @@ using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.UI;
 
-#endregion
+namespace Imperium.Interface.ImperiumUI.Windows.Teleport.Widgets;
 
-namespace Imperium.Interface.ImperiumUI.Windows.Teleportation;
-
-internal class TeleportationWindow : ImperiumWindow
+public class Teleportation : ImpWidget
 {
     private Button tpMainEntrance;
     private Button tpShip;
@@ -31,29 +26,25 @@ internal class TeleportationWindow : ImperiumWindow
     private ImpBinding<float> coordinateY;
     private ImpBinding<float> coordinateZ;
 
-    private Transform content;
-
-    protected override void InitWindow()
+    protected override void InitWidget()
     {
-        content = transform.Find("Content");
-
         tpMainEntrance = ImpButton.Bind(
-            "Presets/MainEntrance", content,
+            "Presets/MainEntrance", transform,
             () => TeleportTo(Imperium.PlayerManager.MainEntranceTPAnchor.Value),
             theme
         );
         tpShip = ImpButton.Bind(
-            "Presets/Ship", content,
+            "Presets/Ship", transform,
             () => TeleportTo(Imperium.PlayerManager.ShipTPAnchor.Value),
             theme
         );
         tpApparatus = ImpButton.Bind(
-            "Presets/Apparatus", content,
+            "Presets/Apparatus", transform,
             () => TeleportTo(Imperium.PlayerManager.ApparatusTPAnchor.Value),
             theme
         );
         ImpButton.Bind(
-            "Presets/Freecam", content,
+            "Presets/Freecam", transform,
             () => TeleportTo(Imperium.Freecam.transform.position),
             theme
         );
@@ -65,13 +56,12 @@ internal class TeleportationWindow : ImperiumWindow
         coordinateY = new ImpBinding<float>(onUpdateSecondary: _ => TeleportToCoords());
         coordinateZ = new ImpBinding<float>(onUpdateSecondary: _ => TeleportToCoords());
 
-        ImpInput.Bind("Coords/CoordsX", content, coordinateX, theme);
-        ImpInput.Bind("Coords/CoordsY", content, coordinateY, theme);
-        ImpInput.Bind("Coords/CoordsZ", content, coordinateZ, theme);
+        ImpInput.Bind("Coords/CoordsX", transform, coordinateX, theme);
+        ImpInput.Bind("Coords/CoordsY", transform, coordinateY, theme);
+        ImpInput.Bind("Coords/CoordsZ", transform, coordinateZ, theme);
 
-        ImpButton.Bind("Buttons/Interactive", content, OnInteractive, theme);
+        ImpButton.Bind("Buttons/Interactive", transform, OnInteractive, theme);
 
-        RegisterWidget<Waypoints>(content, "Waypoints");
 
         Imperium.InputBindings.BaseMap.Teleport.performed += OnInteractiveTeleport;
 
@@ -125,7 +115,7 @@ internal class TeleportationWindow : ImperiumWindow
          * This is for the case the game restricts the player to teleport to the desired coords (e.g. OOB)
          */
         yield return new WaitForSeconds(0.2f);
-        
+
         coordinateX.Set(MathF.Round(Imperium.Player.transform.position.x, 2), invokeSecondary: false);
         coordinateY.Set(MathF.Round(Imperium.Player.transform.position.y, 2), invokeSecondary: false);
         coordinateZ.Set(MathF.Round(Imperium.Player.transform.position.z, 2), invokeSecondary: false);
@@ -133,8 +123,8 @@ internal class TeleportationWindow : ImperiumWindow
 
     private void InitFireExits()
     {
-        fireExitContainer = content.Find("FireExits/ScrollView/Viewport/Content");
-        fireExitsPlaceholder = content.Find("FireExits/Placeholder").GetComponent<TMP_Text>();
+        fireExitContainer = transform.Find("FireExits/ScrollView/Viewport/Content");
+        fireExitsPlaceholder = transform.Find("FireExits/Placeholder").GetComponent<TMP_Text>();
         fireExitTemplate = fireExitContainer.Find("Template").gameObject;
         fireExitTemplate.gameObject.SetActive(false);
 
@@ -146,7 +136,7 @@ internal class TeleportationWindow : ImperiumWindow
     {
         ImpThemeManager.Style(
             themeUpdated,
-            content,
+            transform,
             new StyleOverride("FireExits", Variant.DARKER)
         );